add matomo analytics
this adds tracking for matomo to the docs site. This is accomplished
based on docs for mkdocs material https://squidfunk.github.io/mkdocs-material/setup/setting-up-site-analytics/#custom-site-analytics

note that mkdocs material did not yet have support for this analytics
system until recently, so I updated the dependencies for mkdocs and
mkdocs material as well. There was one other required change for the
social links.

Site is tested to behave same as before.

part of https://github.com/Synthetixio/issues/issues/312